Numpy2 pandas3 (#287)
* add support for numpy 2 and pandas 3

* update versions, docs, fix warnings

* change linter version

* try fix test teardown

* make setup failures resilient

* add setup tools

* add setuptools

* try clear cache

* pin setuptools

* update indentation handling

* try to fix windows permission error

* try fix windows error 2

* try to handle windows test failure

+++ scripts/clean_docs.py
@@ -13,8 +13,9 @@ def clean_docs():
     """Clean all the docs"""
     doc_path = Path(__file__).absolute().parent.parent / "docs"
     # first delete all checkpoints
-    for checkpoint in doc_path.rglob("*.ipynb_checkpoint"):
-        shutil.rmtree(checkpoint)
+    for checkpoint in doc_path.rglob(".ipynb_checkpoints"):
+        if checkpoint.is_dir():
+            shutil.rmtree(checkpoint)
     # make api documentation
     cmd = "jupyter nbconvert --ClearOutputPreprocessor.enabled=True --inplace"
     api_path = doc_path / "api"
@@ -23,7 +24,9 @@ def clean_docs():
         shutil.rmtree(api_path)
     # execute all the notebooks
     for note_book_path in doc_path.rglob("*.ipynb"):
-        if "ipynb_checkpoints" in str(note_book_path):
+        if "_build" in note_book_path.parts or any(
+            "ipynb_checkpoints" in part for part in note_book_path.parts
+        ):
             continue
         result = run(cmd + f" {note_book_path}", shell=True)
         if result.returncode != 0:
